Refactor into current class
